Create/delete NIC as part of the create/delete VM scenario (#113)
Since we have a dynamic number of VMs that we create and delete in a
scenario, we are doing the NIC creation and deletion as part of the bash
scripts. We may want to update the interface in the future to support
this from terraform.

The commit adds function to create and delete a NIC, updates the measure
VM creation/deletion utils to handle NIC creation/deletion as well and
updates the terraform template.