Add terraform for AGC performance testing (#158)
Creating PRs to add scenario for perf testing Application Gateway for
Containers resource. AGC is Azure's new L-7 load balancer built
purposefully for Kubernetes workloads. This scenario will benchmark the
performance as the configuration and cluster size increases.

In this PR, we add the terraform files required to create the AGC
resource along with an AKS cluster. It also deployed the ALB controller
component using helm on the cluster.

Terraform is used to deployed:
* A subnet for AGC resource
* AGC resource and it's child resources like frontend and association
* ALB Controller that is deployed on AKS cluster and sends configuration
to AGC control plane


![image](https://github.com/Azure/telescope/assets/5294363/66074f00-1e20-4a20-8277-3b5f1018dbf1)

![image](https://github.com/Azure/telescope/assets/5294363/47e84a79-9ba2-4a75-b66f-7826bc0b6a00)

![image](https://github.com/Azure/telescope/assets/5294363/f1b701b9-3101-4008-b750-92108631b5bf)