Implement SLO workload (#333)
This pull request introduces several significant changes across multiple
files to enhance the functionality of the `clusterloader2` module and
integrate Kubernetes client capabilities, along with updates to
configuration files and Terraform modules. The key changes include the
addition of a new Kubernetes client class, configuration updates for
deployment churn, and enhancements to the Terraform EKS and AKS modules.

### Enhancements to `clusterloader2`:

*
[`modules/python/clusterloader2/kubernetes_client.py`](diffhunk://#diff-c262bfdbeb316d64561ccf30c6aa0fb6d72f25962edac1008d8666e72ffc7a2cR1-R16):
Introduced a new `KubernetesClient` class to interact with the
Kubernetes API, including methods for describing nodes, listing nodes,
and getting ready nodes.
*
[`modules/python/clusterloader2/slo/slo.py`](diffhunk://#diff-2594888ced631a2698e425dee445068b1b8480a8761e0250ca693e842ddb1d30R1-R147):
Added new functions for overriding and validating the `clusterloader2`
configuration, executing the test, and collecting results.
*
[`modules/python/clusterloader2/utils.py`](diffhunk://#diff-e46631a8b31a54a4e3cb7da1eecbe362f53499f958d8dbd20fb8652e2b0b7d38R4-R26):
Updated the `run_cl2_command` function to support Prometheus and exec
service, and added a new function `get_measurement` to parse measurement
files.
[[1]](diffhunk://#diff-e46631a8b31a54a4e3cb7da1eecbe362f53499f958d8dbd20fb8652e2b0b7d38R4-R26)
[[2]](diffhunk://#diff-e46631a8b31a54a4e3cb7da1eecbe362f53499f958d8dbd20fb8652e2b0b7d38R40-R44)
[[3]](diffhunk://#diff-e46631a8b31a54a4e3cb7da1eecbe362f53499f958d8dbd20fb8652e2b0b7d38R53-R66)

### Configuration updates:

*
[`modules/python/clusterloader2/slo/config/config.yaml`](diffhunk://#diff-490fb4f9566e1a1d18082c232d995bd985221a6ee22f3ee80d8ed6a12db507a1R1-R127):
Added a new SLO configuration for deployment churn with parameters for
nodes, pods, namespaces, and various latency thresholds.
*
[`modules/python/clusterloader2/slo/config/deployment_template.yaml`](diffhunk://#diff-6e902102b9ceb91638fd24417516460d679e78bec901ec817726e47a6258faa3R1-R50):
Added a deployment template with parameters for CPU and memory requests,
and tolerations for node conditions.

### Terraform module updates:

*
[`modules/terraform/aws/eks/main.tf`](diffhunk://#diff-fd5dd7320a4e52fded4d41f24c016f50020f4b1ce1578d83fec5442c437e09ebR71-R72):
Added support for specifying the Kubernetes version in the EKS cluster
configuration.
*
[`modules/terraform/aws/eks/output.tf`](diffhunk://#diff-6424660e540d495fff59cf6b3e4f2fd73a69fd7c59c79218e13e0bec549513f8R5-R11):
Added a new output for EKS dependencies.
*
[`modules/terraform/aws/eks/variables.tf`](diffhunk://#diff-8abbc84758137371997b88238369c9fd9f08384fc96dc0932a3841a66cdf83abR51):
Updated the `eks_config` variable to include an optional
`kubernetes_version`.
*
[`modules/terraform/azure/aks/main.tf`](diffhunk://#diff-4089bc0576048d82e335da51ad0374e2e9c0ae37db3d65c3c784e97871651fd7R54):
Added support for specifying the Kubernetes version and node taints in
the AKS cluster configuration.
[[1]](diffhunk://#diff-4089bc0576048d82e335da51ad0374e2e9c0ae37db3d65c3c784e97871651fd7R54)
[[2]](diffhunk://#diff-4089bc0576048d82e335da51ad0374e2e9c0ae37db3d65c3c784e97871651fd7R70)
*
[`modules/terraform/azure/aks/variables.tf`](diffhunk://#diff-e0b25de6470cd5774403141e8997adee371642381fd8430d6dd467f86d13b59eR73-R79):
Updated the `aks_config` variable to include an optional
`kubernetes_version` and `node_taints`.

### Miscellaneous:

*
[`.yamllint`](diffhunk://#diff-27950a662421192db1847a8f44066c546c6f83964e90d1cd6a30bd02407a2f81L21-R22):
Updated the `ignore` rules to include all YAML and YML files under
`modules/python/clusterloader2`.
*
[`pipelines/perf-eval/slo-nodes100-pods5000.yml`](diffhunk://#diff-d398d62236cc2d7f83da5804c670b55560e2dd772905e29a52ed82ddd031a929R1-R57):
Added a new pipeline configuration for performance evaluation with
specific parameters for AWS and Azure.
*
[`modules/python/requirements.txt`](diffhunk://#diff-28113c509ae49c546b5b3e59a7fea7d86479c5e3662c80835255765cf4d4dd0fR2):
Added the `kubernetes` package to the requirements.