Dependabot for terraform provider updates (#348)
This pull request includes several changes to the Terraform
configuration and Dependabot settings. The most important changes
involve managing dependencies for Terraform and restructuring the
Terraform configuration files.

### Dependabot Configuration:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R1-R15):
Added a new configuration to manage dependencies for Terraform,
specifying the directory, schedule, assignees, reviewers, and open pull
request limit.

### Terraform Configuration:

*
[`modules/terraform/aws/main.tf`](diffhunk://#diff-4fa7e1512b9b19b6cf0ec7166bff84df309a63560e72e26b07a28e5f3fc0a086L18-L27):
Removed the Terraform block that specifies the required version and
providers.
*
[`modules/terraform/aws/versions.tf`](diffhunk://#diff-9c29578937d87994f79600b40d6b6a0219023626d9f4b1f4a045091a2d60b278R1-R9):
Added a new file to specify the Terraform required version and providers
for AWS.
*
[`modules/terraform/azure/main.tf`](diffhunk://#diff-0f7f598ee8901914ae02cf0a5c9c9c1237e1bbe208688856f72a3af4e1cc12bbL37-L46):
Removed the Terraform block that specifies the required version and
providers.
*
[`modules/terraform/azure/versions.tf`](diffhunk://#diff-db80da69c1925b208c7e0de8f495645e88cce4aee7e92412bc60cd5aec7c4a06R1-R9):
Added a new file to specify the Terraform required version and providers
for Azure.