Update configuration for dependabot (#350)
This pull request includes a small change to the
`.github/dependabot.yml` file. The change modifies the configuration for
managing dependencies for Terraform by changing the `directory` key to
`directories` and listing the directory path under it.

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L6-R7):
Changed `directory` key to `directories` and listed the Terraform module
path.This pull request includes a small change to the
`.github/dependabot.yml` file. The change modifies the way directories
are specified for managing Terraform dependencies.

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L6-R7):
Changed from specifying a single directory to using a list of
directories for Terraform dependencies