update pipeline to run aks free vs. standard test (#341)
This pull request includes several updates to the Azure performance
evaluation pipeline and the steps for collecting cloud information and
setting Terraform input variables. The most important changes include
modifying the stages to add new tiers, updating the
`collect-cloud-info.yml` to include SKU tier information, and setting
SKU tier in Terraform input variables.

### Updates to Azure performance evaluation pipeline:

*
[`pipelines/perf-eval/apiserver-benchmark-virtualnodes10-pods100.yml`](diffhunk://#diff-0637ac88f0b2ccd044978f0c20e4c0946867cc47061de1d3febc543e3ae026d3L43-R43):
Renamed the `azure_eastus2` stage to `azure_eastus2_free` and added a
new `azure_eastus2_standard` stage with appropriate parameters.
[[1]](diffhunk://#diff-0637ac88f0b2ccd044978f0c20e4c0946867cc47061de1d3febc543e3ae026d3L43-R43)
[[2]](diffhunk://#diff-0637ac88f0b2ccd044978f0c20e4c0946867cc47061de1d3febc543e3ae026d3R57-R88)

### Collecting cloud information:

*
[`steps/cloud/azure/collect-cloud-info.yml`](diffhunk://#diff-1ede6464d7843d9f0d81b83185843671ba3934477445744a985c8ed31da23c3dR18-R20):
Added the `sku_tier` argument to the cloud info collection script to
capture the SKU tier of the AKS cluster.

### Setting Terraform input variables:

*
[`steps/terraform/set-input-variables-azure.yml`](diffhunk://#diff-7ed544f040598d432bcfbc28c60048961ade121d6ab97091671a37e4d365f86eR68):
Included `SKU_TIER` in the environment variables to be set, which should
be defined by the matrix.