Show warning message if deletion due time is in the past (#398)
Note: this check does not fail the provisioning. It shows the warning
message if the condition is false, and ask to confirm to continue. If
terraform is run with `--auto-approve` the check only shows the warning
message, but doesn't ask to confirm to continue.

Output:
```
Plan: 31 to add, 2 to change, 0 to destroy.
╷
│ Warning: Check block assertion failed
│ 
│   on main.tf line 34, in check "certificate":
│   34:     condition     = timecmp(local.non_computed_tags.deletion_due_time , plantimestamp()) > 0
│     ├────────────────
│     │ local.non_computed_tags.deletion_due_time is "2024-11-12T18:39:54Z"
│ 
│ Deletion due time is in the past: 2024-11-12T18:39:54Z. This might result in the deletion of resources currenty in use. To resolve it, update the creation_time (2024-11-12T16:39:54Z) to current time
```

---------

Co-authored-by: Rafael Mendes Pereira <rafaelme@microsoft.com>