Remove second stages from cilium nodesubnet pipelines to have first stage run without interference (#434)
With the second stage enabled, the same cluster is being used to run
both stages. This needs to be fixed with deeper changes to the
pipelines, so that we can create a separate cluster with a different
run_id to run the second stage. While we work on that, this PR disables
the second stage in both cilium nodesubnet pipelines.