Tune CRI resource consume test case (#470)
This pull request includes several changes aimed at improving the
configuration and validation processes in the `clusterloader2` and
`cri-resource-consume` modules. The most important changes include
adding a new wait step for resource consumption, modifying node count
parameters, and updating validation timeouts.

Configuration updates:

*
[`modules/python/clusterloader2/cri/config/config.yaml`](diffhunk://#diff-c5503c64b4e04dc3563dcb183ba27fd6d8805893f99f195ea3d7f84a9dc7c026R75-R91):
Added a new step to wait for resource consumption by introducing a sleep
measurement with a duration of 1 minute. Also, modified the
`minDesiredNodeCount` to be 80% of the total nodes and set a fixed
timeout of 1 minute for node readiness.

Label adjustments:

*
[`scenarios/perf-eval/cri-resource-consume/terraform-inputs/aws.tfvars`](diffhunk://#diff-d833a32e29f9b21994b6a60a83ddf61bfc0af647420db301f9a7950c65873072L100-R100):
Changed the `agentpool` label from `userpool1` to `userpool0` to reflect
updated resource allocation.

Validation improvements:

*
[`steps/engine/clusterloader2/slo/validate.yml`](diffhunk://#diff-fd071d764ed39deb6feb4c06f9bc7cbd4be1b6ed2a4c1529b886a86875845bf7L12-R18):
Replaced the `OPERATION_TIMEOUT` environment variable with
`VALIDATION_TIMEOUT` to better represent the purpose of the timeout
parameter during node count validation.
*
[`steps/topology/cri-resource-consume/validate-resources.yml`](diffhunk://#diff-2c64573a4f81f1dd3c54c463e3428597b5f0d5cf1e0baad159a3b54d1169c2ebR14-R16):
Added a template reference to the `validate.yml` script and set the
`desired_nodes` parameter to 14 for resource validation.