Created Cilium Network Policy and Cilium Cluster Wide Network Policy Test (#456)
I created a Cilium Network Policy and Cilium Cluster Wide Network Policy
pipeline that will deploy a set number of CCNPS and CNPS that match to
pods. It will also scale to a certain number of nodes and scale to a set
number of pods. The policies have egress and ingress rules to reflect
customer usage and to be able to test control plane resource usage
consumption. This test is created based on the existing service churn
pipeline framework.

I created this test to be able to find the cilium cpu, memory, and
apiserver usage through testing a 1000 node cilium cluster with
endpoints as well as a 3000 node cilium cluster with CES enabled. This
is for the upcoming official release of the cnp/ccnp feature.

Here is the link to my pipeline:
https://dev.azure.com/akstelescope/telescope/_build?definitionId=37 and
here is a link to a successful run:
https://dev.azure.com/akstelescope/telescope/_build/results?buildId=9548&view=results

Since I made changes to files that the other pipelines use, I'm running
Service & Cluster Churn:

Regular Pipelines:
- Service Churn:
https://dev.azure.com/akstelescope/telescope/_build/results?buildId=9545&view=results
- Cluster Churn:
https://dev.azure.com/akstelescope/telescope/_build/results?buildId=9546&view=results