Add support to use custom subnet for AKS CLI module (#484)
This pull request introduces changes to support the optional inclusion
of a subnet ID parameter in the Azure AKS CLI module. The most important
changes include adding a new variable for the subnet ID, updating the
local parameters to conditionally include the subnet ID, and modifying
the resource configurations to use this new parameter.

### Support for optional subnet ID parameter:

*
[`modules/terraform/azure/aks-cli/variables.tf`](diffhunk://#diff-dd28825e80ba149f3e743940a83f0b935bc0fca41dd07c8f9beb10b21a3815b4R19-R30):
Added a new variable `subnet_id` with a default value of `null` and
updated `aks_cli_config` to include an optional `subnet_name`.
*
[`modules/terraform/azure/aks-cli/main.tf`](diffhunk://#diff-1c09e32cd63aa3f4a6cfae577b3db448daedf498df32bd8834fd4344f6b86ab4R39-R46):
Updated the `locals` block to conditionally format the
`subnet_id_paramter` and included it in the `resource "terraform_data"
"aks_cli"` configuration.
[[1]](diffhunk://#diff-1c09e32cd63aa3f4a6cfae577b3db448daedf498df32bd8834fd4344f6b86ab4R39-R46)
[[2]](diffhunk://#diff-1c09e32cd63aa3f4a6cfae577b3db448daedf498df32bd8834fd4344f6b86ab4R103)
*
[`modules/terraform/azure/variables.tf`](diffhunk://#diff-a70d527b3b015f3aa3103f024416bbf9c1e2828c82890a61a97316c23e5402efR210):
Updated the `aks_cli_config_list` variable to include an optional
`subnet_name`.
*
[`modules/terraform/azure/main.tf`](diffhunk://#diff-0f7f598ee8901914ae02cf0a5c9c9c1237e1bbe208688856f72a3af4e1cc12bbR98):
Modified the `module "aks-cli"` block to use the `subnet_id` parameter
by trying to fetch it from the local subnets.