Add ena express block in EKS launch template only if ena_express is not null (#594)
This pull request includes several changes to the AWS EKS module in
Terraform. The primary focus is on modifying the handling of the
`ena_express` parameter and updating related configurations and tests.

**What is different from previous PR: the default ena_express is null
and it add the ena block in the launch template only if ena_express is
not null.** See commit:
https://github.com/Azure/telescope/pull/594/commits/846ab9e3c5e97841fd05985d9c640697a0c2276a

Changes to AWS EKS module:

*
[`modules/terraform/aws/eks/main.tf`](diffhunk://#diff-fd5dd7320a4e52fded4d41f24c016f50020f4b1ce1578d83fec5442c437e09ebL188-R197):
Updated the `ena_srd_specification` block to use a dynamic block based
on the `ena_express` variable.
*
[`modules/terraform/aws/eks/variables.tf`](diffhunk://#diff-8abbc84758137371997b88238369c9fd9f08384fc96dc0932a3841a66cdf83abL53-R53):
Changed the default value of the `ena_express` variable from `false` to
`null` in both `eks_config` and `eks_config_list` variables.
[[1]](diffhunk://#diff-8abbc84758137371997b88238369c9fd9f08384fc96dc0932a3841a66cdf83abL53-R53)
[[2]](diffhunk://#diff-9bd8d786978d8f7797b6b7bb3cf1d957ceab3be62c9fb17ec3ad6a3ee53f22caL122-R122)

Changes to tests:

*
[`modules/terraform/aws/tests/test_eks_node_groups.tftest.hcl`](diffhunk://#diff-d510eb2a3a55b7e7d3c553c9fca75fc15dd64d3e9be6c8497c06db64aefca94dL111-L121):
Updated test conditions to check for the absence of
`ena_srd_specification` when `ena_express` is `null`.

Other changes:

*
[`scenarios/perf-eval/nap-c4n10p100/terraform-inputs/aws.tfvars`](diffhunk://#diff-3b945e4bddb2cabffe551d39e1ef108c6664d72b2eeddf8b5f0f6655b11a35a5L3-R3):
Increased the `deletion_delay` from `2h` to `72h`.

---------

Co-authored-by: Rafael Mendes Pereira <rafaelme@microsoft.com>