Skip terraform collect step if the resources are not provisioned (#603)
This pull request refactors the resource cleanup logic in Azure
pipelines by improving the handling of conditions for resource
management. The changes simplify and centralize the condition checks to
ensure better maintainability and readability.

### Improvements to condition handling:

* **`steps/cleanup-resources.yml`:** Refactored the condition for
including the `collect-terraform-operation-metadata.yml` template to
remove redundancy. The `condition` parameter is now passed explicitly to
the template, ensuring that the `SKIP_RESOURCE_MANAGEMENT` variable is
consistently evaluated.

* **`steps/collect-terraform-operation-metadata.yml`:** Added a new
`condition` parameter (defaulting to `true`) to allow conditional
execution of the steps within the template. This centralizes the
condition logic and ensures steps are only executed when the condition
evaluates to `true`.