Skip terraform collect step when SKIP_RESOURCE_MANAGEMENT is true (#619)
This pull request includes updates to the resource cleanup and metadata
collection processes in the pipeline. The key changes involve
simplifying conditional logic, adding a new condition for metadata
collection, and minor structural adjustments to the YAML files.

### Changes to resource cleanup logic:
*
[`steps/cleanup-resources.yml`](diffhunk://#diff-0baba20632afdc0472aed05c703778b9f428731dc8fedac4169d47ea05adf245L35-R35):
Simplified the conditional logic by removing the redundant check for
`SKIP_RESOURCE_MANAGEMENT` in the `if` statement for the
`collect-terraform-operation-metadata.yml` template.

### Changes to metadata collection:
*
[`steps/collect-terraform-operation-metadata.yml`](diffhunk://#diff-5b2db3ca18632b93bdbbc7b4dfb4c85a86b8a27d8fa61415ac56a37e082746cbR16):
Added a condition to ensure that the "Collect Terraform Operation
Metadata" step only runs if `SKIP_RESOURCE_MANAGEMENT` is not set to
`true`.

### Structural adjustments:
*
[`steps/collect-terraform-operation-metadata.yml`](diffhunk://#diff-5b2db3ca18632b93bdbbc7b4dfb4c85a86b8a27d8fa61415ac56a37e082746cbR7):
Added a blank line for better readability in the `parameters:` section.