Skip deleting resource for debugging (#668)
When we debug a new pipeline and it fails, the cluster is deleted before
we have a chance to investigate.
This PR adds a variable "SKIP_RESOURCE_DELETION". When set, it the
pipeline to skips the deletion of resources in manual runs.
The owner of a pipeline run is responsible to delete the resource
manually when it's skipped.

To use it, one can set the variable in the pipeline yaml.

```
variables:
    SKIP_RESOURCE_DELETION: true
```

This is tested in
https://dev.azure.com/akstelescope/telescope/_build/results?buildId=25279&view=logs&j=de223fca-0b0d-5d04-db32-a8c4d0bcfd88&t=c58ccdb9-3032-53d1-f00a-a15c1c92ae2e

---------

Co-authored-by: Lei Yao <yaolei@microsoft.com>