Deploy Metrics Server addon deployment with manifest for EKS Auto Mode (#759)
Deploy Metrics Server addon deployment with manifest for EKS Auto Mode
to prevent deployment issues


### Enhancements to Metrics Server Addon Handling:
- **Manifest-based Deployment**: Added a `terraform_data` resource to
deploy the Metrics Server addon using a Kubernetes manifest when EKS
Auto Mode is enabled but the addon is not explicitly configured. This
ensures faster readiness compared to the official EKS addon.
- **Explicit Addon Configuration**: Updated logic to exclude the Metrics
Server addon from the `_eks_addons_map` when deploying via manifest,
preventing conflicts.

### Improvements to Addon Configuration:
- **Dynamic Timeouts**: Introduced dynamic timeout settings for addon
creation, update, and deletion, allowing greater flexibility in resource
operations.
- **Conflict Resolution**: Added `resolve_conflicts_on_create` with the
"OVERWRITE" strategy to handle potential conflicts during addon
creation.

### Updates to Unit Tests:
- **Validation of Manifest Deployment**: Added test cases to ensure
Metrics Server is deployed via manifest when Auto Mode is enabled and
the addon is not explicitly configured, and that it is not deployed when
Auto Mode is disabled.
[[1]](diffhunk://#diff-90ba4574efdc1b59287ec73a6fba7282bc726d7c3358ddebd2b373669f1833b1L105-R126)
[[2]](diffhunk://#diff-90ba4574efdc1b59287ec73a6fba7282bc726d7c3358ddebd2b373669f1833b1R179-R184)
- **Explicit Addon Testing**: Added tests to confirm the Metrics Server
addon is created when explicitly configured in `eks_addons`, and not
deployed via manifest in such cases.

### Outputs for Unit Testing:
- **New Output**: Added `apply_metrics_server_addon` to expose the
manifest-based Metrics Server deployment resource for unit tests.

---------

Co-authored-by: Rafael Mendes Pereira <rafaelme@microsoft.com>