fix(job-controller): fix job scheduling error, attempting to create flowschema error (#937)
Job Scheduling Pipeline is failing: KWOK is trying to create a
FlowSchema that references the "exempt" PriorityLevelConfiguration,
which is not allowed in Kubernetes (on EKS). Fail on eks but passing for
aks

Pipeline:
https://dev.azure.com/akstelescope/telescope/_build?definitionId=55&_a=summary
The issue is that the KWOK manifests contain a FlowSchema that
references the "exempt" PriorityLevelConfiguration, which is protected
in Kubernetes. This solution is to skip applying FlowSchema resources
that exempt PriorityLevelConfiguration

Co-authored-by: vittoria salim <vslaim@microsoft.com>